refactor text example structure
