+++ doc_swift_ui_gallery/swift_ui_gallery/swift_ui_gallery/primitives/text/localization/TextLocalizationPage.swift
@@ -1,0 +1,13 @@
+//
+
+import SwiftUI
+
+struct TextLocalizationPage: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TextLocalizationPage()
+}
